I got the functionality to update the Json file, save the image locally, and retrieve it to show on the page.
Added links now work.
The only problem is react-router-dom not working. It worked for a moment, and then I added the formData.link functionality and somewhere in there I screwed it up.
I had a more hacky way of making it work before react router, and I'm loath to switch back but if I cant get it working in an hour I will so I can move on to the next site.